RAG/requirements.txt
python-dotenv>=1.0.0
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.4
llama-index-readers-file>=0.1.4
llama-index-llms-together>=0.1.3
llama-index-vector-stores-chroma>=0.1.1
sentence-transformers>=2.2.2
numpy>=1.24.0
requests>=2.31.0
pypdf>=4.0.0
docx2txt>=0.8
pillow>=10.0.0
chromadb>=0.4.18